requirements.txt
astroid==2.3.3
autopep8==1.5
certifi==2020.4.5.1
chardet==3.0.4
cycler==0.10.0
idna==2.9
isort==4.3.21
joblib==0.14.1
kaggle==1.5.6
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
matplotlib==3.2.1
mccabe==0.6.1
numpy==1.18.2
pandas==1.0.3
Pillow==7.1.1
pycodestyle==2.5.0
pylint==2.4.4
pyparsing==2.4.7
python-dateutil==2.8.1
python-slugify==4.0.0
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.23.0
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
sklearn==0.0
text-unidecode==1.3
tqdm==4.45.0
urllib3==1.24.3
utils==1.0.1
wrapt==1.11.2
